fix(tests): avoid redundant cast in SCP cleanup diagnostics
Cast each selected enum value to unsigned int so the diagnostic keeps the
correct printf argument type and passes readability-redundant-casting.
This resolves the clang-tidy failures in the SCP cleanup regression test.

+++ tests/test_scp_nonblock.c
@@ -243,8 +243,8 @@ static int test_error_cleanup_eagain(int send_case)
                 send_case ? "send" : "recv", (void *)result, rc,
                 session->err_msg ? session->err_msg : "(null)", api_calls,
                 eagain_returns, ctx.recv_calls, ctx.send_calls,
-                (unsigned int)(send_case ? session->scpSend_state :
-                                           session->scpRecv_state),
+                send_case ? (unsigned int)session->scpSend_state :
+                            (unsigned int)session->scpRecv_state,
                 (void *)(send_case ? session->scpSend_channel :
                                     session->scpRecv_channel));
         failed = 1;
